util: Extract _Snmp from eaton.py to util/snmp.py
With new Raritan power driver planned, which will be using SNMP,
repeating the same wrapper code for pySNMP would be unwise.
Instead, we shall extract the wrapper code from eaton.py to
utils and rename the class to SimpleSNMP.
